2-get_bit.c

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -7,7 +7,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	int value;
+	unsigned long int value;
 
 	if (n == 0 || index == 0)
 		exit(1);
